sistema-solar/requirements.txt
numpy==1.24.4
pygame==2.6.1
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10